requirements.txt
moviepy==1.0.3
openai-whisper==20230314
torch==2.0.1
transformers==4.30.2
loguru==0.7.0
pytest==7.3.1